+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 8.6
 - Add (Student) Widgets class in classes/core/Widgets.php
 - Add (Student) Widget interface and individual Widget classes in classes/core/Widget.php
 - Use RosarioSIS\Widgets in Widgets.fnc.php
+- Add StaffWidgets class in classes/core/StaffWidgets.php
 
 Changes in 8.5
 --------------